// apps/web/src/nodes/component-types.generated.ts
export const COMPONENT_TYPES = [
  "Button",
  "Calculate",
  "Compare",
  "Constant",
  "Counter",
  "Delay",
  "Figma",
  "Force",
  "Function",
  "Gate",
  "HallEffect",
  "Hotkey",
  "I2cDevice",
  "Interval",
  "Ldr",
  "Led",
  "Llm",
  "Matrix",
  "Midi",
  "Monitor",
  "Motion",
  "Mqtt",
  "Music",
  "Note",
  "Oscillator",
  "Piezo",
  "Pixel",
  "Pn532",
  "Potentiometer",
  "Proximity",
  "RangeMap",
  "Relay",
  "Rgb",
  "Sensor",
  "Servo",
  "Smooth",
  "Stepper",
  "Switch",
  "Tilt",
  "Trigger",
  "Vibration",
] as const;

export type ComponentType = (typeof COMPONENT_TYPES)[number];